Time: 1 ms (46.59%) | Memory: 62.8 MB (73.92%) - LeetSync

+++ 11-container-with-most-water/container-with-most-water.cpp
@@ -2,10 +2,10 @@ class Solution {
 public:
     int maxArea(vector<int>& height) {
         int i=0,j=height.size()-1;
-        long long res=0;
+        int res=0;
         while(i<=j){
-            long long  w=j-i;
-            long long l=min(height[i],height[j]);
+            int  w=j-i;
+            int l=min(height[i],height[j]);
             res=max(res,l*w);
             if(height[i]<=height[j]){
                 i++;
